backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.28.0
greenlet>=3.0.0
authlib==1.2.0
httpx==0.24.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
alembic==1.12.0
pydantic==2.7.1
pydantic-settings>=2.0.0
email-validator>=2.0.0
boto3>=1.34.0
python-multipart>=0.0.6
